flask_app/requirements.txt
celery==4.4.2
Flask==1.1.2
Flask-Caching==1.8.0
Flask-Cors==3.0.8
flask-fs==0.6.1
Flask-JWT-Extended==3.24.1
Flask-Mail==0.9.1
flask-mongoengine==0.9.5
Flask-RBAC==0.2.1
flask-redis==0.4.0
flask-restx==0.2.0
Flask-Script==2.0.6
gevent==20.5.0
greenlet==0.4.15
gunicorn==20.0.4
mongoengine==0.20.0
Pillow==7.1.2
python-dotenv==0.13.0
Werkzeug==0.16.1